project/grocify/tests/test_models.py
"""
Critical model tests for business logic validation.
"""
import pytest
from decimal import Decimal
from django.test import TestCase
from django.core.exceptions import ValidationError

from .factories import (
    ProductFactory, CustomerFactory, CreditSaleFactory, 
    LoyaltyProfileFactory, LoyaltyTierFactory, InventoryItemFactory,
    SaleTransactionFactory, LocationFactory
)
from customers.models import Customer
from credit.models import CreditSale, CreditPayment
from loyalty.models import LoyaltyProfile
from inventory.models import Product


class ProductModelTest(TestCase):
    """Test Product model business logic."""

    def test_profit_margin_calculation(self):
        """Test profit margin is calculated correctly."""
        product = ProductFactory(purchase_price=Decimal('10.00'), selling_price=Decimal('15.00'))
        expected_margin = ((Decimal('15.00') - Decimal('10.00')) / Decimal('10.00')) * 100
        self.assertEqual(product.profit_margin, expected_margin)

    def test_selling_price_with_tax(self):
        """Test selling price with tax calculation."""
        product = ProductFactory(selling_price=Decimal('100.00'), tax_percent=Decimal('10.0'))
        expected_price_with_tax = Decimal('100.00') * (1 + Decimal('10.0') / 100)
        self.assertEqual(product.selling_price_with_tax, expected_price_with_tax)

    def test_negative_price_validation(self):
        """Test that negative prices are not allowed."""
        with self.assertRaises(ValidationError):
            product = ProductFactory(purchase_price=Decimal('-1.00'))
            product.full_clean()

    def test_sku_uniqueness(self):
        """Test SKU uniqueness constraint."""
        ProductFactory(sku='TEST123')
        with self.assertRaises(Exception):  # IntegrityError
            ProductFactory(sku='TEST123')


class CustomerModelTest(TestCase):
    """Test Customer model business logic."""

    def test_outstanding_balance_calculation(self):
        """Test outstanding balance calculation from credit sales."""
        customer = CustomerFactory()
        
        # Create credit sales
        CreditSaleFactory(customer=customer, credit_amount=Decimal('1000.00'))
        CreditSaleFactory(customer=customer, credit_amount=Decimal('500.00'))
        
        customer.recalculate_outstanding()
        self.assertEqual(customer.outstanding_balance, Decimal('1500.00'))

    def test_walkin_customer_creation(self):
        """Test walk-in customer singleton pattern."""
        customer1 = Customer.get_walkin_customer()
        customer2 = Customer.get_walkin_customer()
        
        self.assertEqual(customer1.id, customer2.id)
        self.assertEqual(customer1.phone, '0000000000')
        self.assertEqual(customer1.name, 'Walk-In Customer')

    def test_phone_validation(self):
        """Test phone number format validation."""
        with self.assertRaises(ValidationError):
            customer = CustomerFactory(phone='invalid-phone')
            customer.full_clean()


class CreditSaleModelTest(TestCase):
    """Test CreditSale model business logic."""

    def test_balance_due_calculation(self):
        """Test balance due calculation with payments."""
        credit_sale = CreditSaleFactory(credit_amount=Decimal('1000.00'))
        
        # Add payments
        CreditPayment.objects.create(
            credit_sale=credit_sale,
            amount_paid=Decimal('300.00'),
            payment_date=credit_sale.issued_on.date()
        )
        CreditPayment.objects.create(
            credit_sale=credit_sale,
            amount_paid=Decimal('200.00'),
            payment_date=credit_sale.issued_on.date()
        )
        
        self.assertEqual(credit_sale.balance_due, Decimal('500.00'))

    def test_overpayment_prevention(self):
        """Test that overpayments are prevented."""
        credit_sale = CreditSaleFactory(credit_amount=Decimal('100.00'))
        
        with self.assertRaises(ValidationError):
            payment = CreditPayment(
                credit_sale=credit_sale,
                amount_paid=Decimal('150.00'),  # More than balance
                payment_date=credit_sale.issued_on.date()
            )
            payment.full_clean()

    def test_settlement_logic(self):
        """Test automatic settlement when fully paid."""
        credit_sale = CreditSaleFactory(credit_amount=Decimal('100.00'))
        
        CreditPayment.objects.create(
            credit_sale=credit_sale,
            amount_paid=Decimal('100.00'),
            payment_date=credit_sale.issued_on.date()
        )
        
        credit_sale.refresh_from_db()
        self.assertTrue(credit_sale.is_settled)


class LoyaltyProfileModelTest(TestCase):
    """Test LoyaltyProfile model business logic."""

    def test_tier_assignment_by_points(self):
        """Test automatic tier assignment based on points."""
        bronze_tier = LoyaltyTierFactory(name='Bronze', minimum_points=0)
        silver_tier = LoyaltyTierFactory(name='Silver', minimum_points=100)
        gold_tier = LoyaltyTierFactory(name='Gold', minimum_points=500)
        
        profile = LoyaltyProfileFactory(points=250)
        profile.assign_tier()
        
        self.assertEqual(profile.tier, silver_tier)

    def test_points_multiplier_calculation(self):
        """Test points calculation with tier multipliers."""
        tier = LoyaltyTierFactory(name='Gold', multiplier=Decimal('1.5'))
        profile = LoyaltyProfileFactory(tier=tier)
        
        # Assuming base rate is 5% and tier multiplier is 1.5
        # 100 * 0.05 * 1.5 = 7.5 points (rounded to 7)
        from loyalty.utils import calculate_earned_points
        points = calculate_earned_points(100.0, 'Gold')
        
        # This will depend on your actual implementation
        self.assertGreater(points, 5)  # Should be more than base rate


class InventoryModelTest(TestCase):
    """Test inventory-related model business logic."""

    def test_stock_deduction_accuracy(self):
        """Test accurate stock deduction during sales."""
        location = LocationFactory()
        inventory = InventoryItemFactory(qty_on_hand=100, location=location)
        
        # Simulate a sale
        original_qty = inventory.qty_on_hand
        sale_qty = 10
        
        inventory.qty_on_hand -= sale_qty
        inventory.save()
        
        inventory.refresh_from_db()
        self.assertEqual(inventory.qty_on_hand, original_qty - sale_qty)

    def test_negative_stock_prevention(self):
        """Test prevention of negative stock levels."""
        inventory = InventoryItemFactory(qty_on_hand=5)
        
        # This should be prevented at the business logic level
        # The actual implementation depends on your business rules
        with self.assertRaises(ValueError):
            if inventory.qty_on_hand < 10:
                raise ValueError("Insufficient stock")


class SaleTransactionModelTest(TestCase):
    """Test SaleTransaction model business logic."""

    def test_invoice_number_generation(self):
        """Test unique invoice number generation."""
        sale1 = SaleTransactionFactory()
        sale2 = SaleTransactionFactory()
        
        self.assertNotEqual(sale1.invoice_number, sale2.invoice_number)
        self.assertTrue(sale1.invoice_number.startswith('INV-'))

    def test_change_calculation(self):
        """Test change due calculation."""
        sale = SaleTransactionFactory(
            total_amount=Decimal('95.00'),
            amount_paid=Decimal('100.00')
        )
        sale.change_due = sale.amount_paid - sale.total_amount
        
        self.assertEqual(sale.change_due, Decimal('5.00'))

    def test_points_earned_calculation(self):
        """Test automatic points calculation on sale."""
        customer = CustomerFactory()
        profile = LoyaltyProfileFactory(customer=customer)
        
        sale = SaleTransactionFactory(
            customer=customer,
            total_amount=Decimal('200.00')
        )
        
        # Points should be calculated automatically
        # This depends on your signal implementation
        self.assertGreaterEqual(sale.points_earned, 0)