requirements.txt
# Core Django Framework
Django>=4.2,<5.0
django-crispy-forms>=2.0
crispy-bootstrap5>=0.7

# Database
psycopg2-binary>=2.9.5

# Image handling
Pillow>=10.0.0

# Barcode generation
treepoem>=3.25.0

# Development and Testing
pytest-django>=4.5.0
factory-boy>=3.2.0
coverage>=7.0.0
pytest-cov>=4.0.0

# Security and Environment
python-decouple>=3.6

# Optional: Production server
gunicorn>=21.0.0

# Optional: Static file handling
whitenoise>=6.5.0

# Optional: Debug toolbar for development
django-debug-toolbar>=4.0.0